refactor: move Footer component to own folder
